backend/requirements.txt
# FastAPI核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# 用户认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# HTTP客户端
httpx==0.26.0

# 定时任务
apscheduler==3.10.4

# 数据加密
pycryptodome==3.20.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 电商平台SDK（需要根据实际情况安装）
# taobao-sdk
# jd-sdk
# pdd-sdk
